Refactor search control implementation
