+++ client/src/components/map/controls/Search/Search.css
@@ -1,0 +1,17 @@
+.search-button {
+    width: 40px;
+    height: 40px;
+    background-color: #121212;
+    border: none;
+    cursor: pointer;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 0;
+    border-radius: 0;
+}
+
+.search-icon {
+    font-size: 20px;
+    color: white;
+}
